Fixed a couple bugs
There was a bug where the container dims were not being sorted in the
recursive func. This would essentially disable the opportunity to
rotate a box across any of the three axes to fit in the container
[within].

Further, added a catch in the recursive func in case no box would fit
in the container — volume array would all be zero. Before, the max
would be zero, and then the “for i in 0..<boxesVolume.count {
            if (maxVolume == boxesVolume[i])” would then match on the
very first index. Thus, a box would be consumed even though it did not
fit. Now, if the max volume is zero in the array, we return the
original boxes to Step 1 of the algorithm.